+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 #### :bug: Bug fix
 
 - Preserve multibyte characters when wrapping long source lines in compiler code frames. https://github.com/rescript-lang/rescript/pull/8520
+- Fix reanalyze optional-argument diagnostics for functions passed or returned as first-class values. https://github.com/rescript-lang/rescript/pull/8321
 
 #### :memo: Documentation